+++ tests/fibonacci_call.cpp
@@ -7,12 +7,12 @@ int main() {
     interpretator._frames.push(start_frame);
     interpretator._current_frame = start_frame;
 
-    uint8_t *start_bytecode = reinterpret_cast<uint8_t *>(calloc(1000, 1));
+    int64_t *start_bytecode = reinterpret_cast<int64_t *>(calloc(1000, 1));
     interpretator.SetPtr(start_bytecode, 1000, start_bytecode);
     // r0 = 5
     interpretator.create_lmov(10, 0);
     // call fibonacci(r0)
-    std::vector<uint8_t> v;
+    std::vector<int64_t> v;
     v.push_back(0);
     interpretator.create_call(0, 1, v);
     // r0 = _return_value
@@ -20,10 +20,10 @@ int main() {
     interpretator.create_exit(0);
 
     
-    uint64_t func_id = interpretator.CreateFunction("fibonacci", 1 /* n_args */);
+    int64_t func_id = interpretator.CreateFunction("fibonacci", 1 /* n_args */);
     Function *function = interpretator.GetFuncById(func_id);
     function->_n_regs = 6;
-    function->_bytecode = (uint8_t *) calloc(10000, 1);
+    function->_bytecode = (int64_t *) calloc(10000, 1);
     function->_bytecode_len = 10000;
     interpretator.SetPtr(function->_bytecode, function->_bytecode_len, function->_bytecode);
 
@@ -39,14 +39,14 @@ int main() {
     // r3 = r0 - r1
     interpretator.create_lsub(0, 1, 3);
     // call fibonacci(r3)
-    std::vector<uint8_t> v1; v1.push_back(3);
+    std::vector<int64_t> v1; v1.push_back(3);
     interpretator.create_call(0, 1, v1);
     // r3 = return value
     interpretator.create_lmov_return(3);
     // r4 = r0 - r2
     interpretator.create_lsub(0, 2, 4);
     // call fibonacci(r3)
-    std::vector<uint8_t> v2; v2.push_back(4);
+    std::vector<int64_t> v2; v2.push_back(4);
     interpretator.create_call(0, 1, v2);
     // r4 = return value
     interpretator.create_lmov_return(4);
@@ -62,7 +62,6 @@ int main() {
     interpretator.Dump();
     std::cout << "------------------------" << std::endl;
 
-    // interpretator.SetPtr(function->_bytecode, function->_bytecode_len, function->_bytecode);
     interpretator.SetPtr(start_bytecode, 1000, start_bytecode);
     interpretator.Run(true);
 }
